requirements.txt
moviepy>=1.0.3
pillow>=9.0.0
openai>=1.3.0
python-dotenv>=0.19.0
tqdm>=4.65.0